+++ show_gnmi_util/formatter/__init__.py
@@ -4,6 +4,6 @@ Formatters for different show commands
 
 from .show_version_formatter import ShowVersionFormatter
 from .show_vlan_brief_formatter import ShowVlanBriefFormatter
-from .show_clock import ShowClockFormatter, ShowClockTimezonesFormatter
+from .show_clock_formatter import ShowClockFormatter, ShowClockTimezonesFormatter
 
 __all__ = ['ShowVersionFormatter', 'ShowVlanBriefFormatter', 'ShowClockFormatter', 'ShowClockTimezonesFormatter']
