+++ show_gnmi_util/scripts/show_clock_sample.json
