+++ show_gnmi_util/scripts/show_clock_timezones_sample.json
